requirements.txt
requests
beautifulsoup4
requests_html
aiohttp
asyncio
PySocks